# Speed up ListIndexableRepos query for uncloned